scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
instaloader>=4.10.0
tweepy>=4.14.0
python-dotenv>=1.0.0
mysql-connector-python>=8.2.0
PyMySQL>=1.1.0
lxml>=4.9.0
fake-useragent>=1.4.0